+++ grout/migrations/0009_auto_20150429_1627.py
@@ -14,11 +14,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='itemschema',
             name='next_version',
-            field=models.OneToOneField(related_name='previous_version', null=True, editable=False, to='grout.ItemSchema'),
+            field=models.OneToOneField(related_name='previous_version', null=True, editable=False, to='grout.ItemSchema', on_delete=models.CASCADE),
         ),
         migrations.AlterField(
             model_name='recordschema',
             name='next_version',
-            field=models.OneToOneField(related_name='previous_version', null=True, editable=False, to='grout.RecordSchema'),
+            field=models.OneToOneField(related_name='previous_version', null=True, editable=False, to='grout.RecordSchema', on_delete=models.CASCADE),
         ),
     ]